+++ examples/mutag_gin.py
@@ -20,8 +20,9 @@ parser.add_argument('--wandb', action='store_true', help='Track experiment')
 args = parser.parse_args()
 
 device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-init_wandb(name=f'GIN-{args.dataset}', lr=args.lr, epochs=args.epochs,
-           hidden_channels=args.hidden_channels, device=device)
+init_wandb(name=f'GIN-{args.dataset}', batch_size=args.batch_size, lr=args.lr,
+           epochs=args.epochs, hidden_channels=args.hidden_channels,
+           num_layers=args.num_layers, device=device)
 
 path = osp.join(osp.dirname(osp.realpath(__file__)), '..', 'data', 'TU')
 dataset = TUDataset(path, name=args.dataset).shuffle()